backend/requirements.txt
neo4j
python-dotenv
tavily-python
fastapi==0.115.0
uvicorn==0.30.6